Fix potential busy loop with abnormal poll
In this situation, poll is returning instantly with len=1 but
none of the fds have any revents set. This results in a busy loop
causing max cpu core usage.

This can be fixed by checking if the fds did nothing and sleeping
for the equivalent amount that poll would have normally took if
no fds got any events and returned len=0.